consolidate snap-model-component into tag-model-component
